feat(apigen): 6035e728 unify recordsrv priority/weight/port + nested weight/activeUsersCount Go types to *uint32
Changed 5 record_srv.go catalog fields from goTypeInt64 to goTypeUint32,
matching the SDK's actual *uint32 (or uint32-valued) wire type:

  - Priority, Weight, Port (top-level, Required, both-scope) — SDK:
    RecordSRV.Priority/.Weight/.Port *uint32.
  - SRVRecordAwsRte53RecordInfo.Weight (nested, response-only) — SDK:
    Awsrte53recordinfo.Weight uint32.
  - SRVRecordMsAdUserData.ActiveUsersCount (nested, response-only) — SDK:
    MsserverAduserData.ActiveUsersCount uint32.

Audited the two remaining goTypeInt64 fields in record_srv.go
(CreationTime, LastQueried) against the SDK and confirmed they are
correctly int64: both are *ibclient.UnixTime, a time.Time wrapper that
(un)marshals as a signed Unix epoch-seconds integer, not a 32-bit
unsigned quantity. Documented the audit result in the file's doc comment.

Regenerated apis/{common,cluster,namespaced}/recordsrv and both recordsrv
CRDs (cluster + namespaced) via the openapi2crd generator + make generate.
All 5 fields now emit format: int32. Diffed CRD marker counts against
origin/main to confirm nothing else moved: minimum/maximum unchanged
(1/1, the TTL markers), format: int64 count dropped from 12 to 4 (the 8
occurrences of the 5 converted fields across ForProvider/AtProvider,
minus the 2 unrelated boilerplate observedGeneration fields already
counted), format: int32 rose from 2 to 10.

Fixed README.md L1753: DNSView Upgrade Notes numeral corrected from
'thirteen' to 'seventeen' — the list below it already enumerated 17
fields (8 top-level + 3 responseRateLimiting + 6 scavengingSchedule),
widened by c6909220's audit but the numeral was never updated. Added a
matching SRVRecord integer field type unification Upgrade Notes section
documenting the CRD schema change and required CRD reapply.

Re-ran the V.C52 use-flag-gate verify block by hand: fails=0 — untouched.

go build ./apis/... passes standalone. go build ./... (and therefore make
reviewable) fails only inside internal/controller/recordsrv, which still
plumbs priority/weight/port through *int64 via uint32OrZero/
uint32PtrToInt64Ptr/uint32PtrOrZero and now has a type mismatch at every
call site — the same intermediate state c6909220 left before
IN-FIX-DNSVIEW-TTL-CTRL, and out of apigen's charter (internal/controller/**
belongs to the controller ant). Filed controller follow-up ticket
e8bbac98-cd21-4781-ad69-daea8c2e2d02 on this branch and added it as a
dependency of the existing review ticket 00768ec0.

+++ apis/common/recordsrv/recordsrv_types.go
@@ -33,7 +33,7 @@ type SRVRecordAwsRte53RecordInfo struct {
 	// Type of Amazon Route 53 resource record.
 	Type *string `json:"type,omitempty"`
 	// Value that determines the portion of traffic for this record in weighted routing. The range is from 0 to 255.
-	Weight *int64 `json:"weight,omitempty"`
+	Weight *uint32 `json:"weight,omitempty"`
 }
 
 // SRVRecordCloudInfo carries Cloud API delegation/ownership information for a cloud-managed SRVRecord (mirrors the SDK's GridCloudapiInfo struct).
@@ -69,7 +69,7 @@ type SRVRecordCloudInfoDelegatedMember struct {
 // SRVRecordMsAdUserData carries Microsoft Active Directory user information for an MS-managed SRVRecord (mirrors the SDK's MsserverAduserData struct).
 type SRVRecordMsAdUserData struct {
 	// The number of active users.
-	ActiveUsersCount *int64 `json:"activeUsersCount,omitempty"`
+	ActiveUsersCount *uint32 `json:"activeUsersCount,omitempty"`
 }
 
 // SRVRecordParameters are the configurable fields of a SRVRecord.
@@ -82,13 +82,13 @@ type SRVRecordParameters struct {
 	Target *string `json:"target"`
 	// Priority of the record (0-65535) — lower values are preferred. Changing this changes the record's _ref.
 	// +kubebuilder:validation:Required
-	Priority *int64 `json:"priority"`
+	Priority *uint32 `json:"priority"`
 	// Relative weight (0-65535) for records with the same priority. Changing this changes the record's _ref.
 	// +kubebuilder:validation:Required
-	Weight *int64 `json:"weight"`
+	Weight *uint32 `json:"weight"`
 	// TCP/UDP port (0-65535) on the target host. Changing this changes the record's _ref.
 	// +kubebuilder:validation:Required
-	Port *int64 `json:"port"`
+	Port *uint32 `json:"port"`
 	// Comment for the record; maximum 256 characters.
 	Comment *string `json:"comment,omitempty"`
 	// Time-to-live in seconds. Zero means the record is not cached. Must be non-negative (0-2147483647); to inherit the zone/grid default, set useTtl to false rather than passing a negative sentinel value.
@@ -122,11 +122,11 @@ type SRVRecordObservation struct {
 	// Target host in FQDN format. Changing this changes the record's _ref.
 	Target *string `json:"target,omitempty"` // atProvider
 	// Priority of the record (0-65535) — lower values are preferred. Changing this changes the record's _ref.
-	Priority *int64 `json:"priority,omitempty"` // atProvider
+	Priority *uint32 `json:"priority,omitempty"` // atProvider
 	// Relative weight (0-65535) for records with the same priority. Changing this changes the record's _ref.
-	Weight *int64 `json:"weight,omitempty"` // atProvider
+	Weight *uint32 `json:"weight,omitempty"` // atProvider
 	// TCP/UDP port (0-65535) on the target host. Changing this changes the record's _ref.
-	Port *int64 `json:"port,omitempty"` // atProvider
+	Port *uint32 `json:"port,omitempty"` // atProvider
 	// Comment for the record; maximum 256 characters.
 	Comment *string `json:"comment,omitempty"` // atProvider
 	// Time-to-live in seconds. Zero means the record is not cached. Must be non-negative (0-2147483647); to inherit the zone/grid default, set useTtl to false rather than passing a negative sentinel value.
